+++ src/quo/__init__.py
@@ -92,4 +92,4 @@ from quo.i_o import (
 
 from quo.shortcuts import container
 
-__version__ = "2022.2.2"
+__version__ = "2022.3"