+++ src/quo/prompt.py
@@ -1459,3 +1459,14 @@ def create_confirm_session(
     return session
 
 
+
+
+hidden = [True]  # Nonlocal
+kb = KeyBinder()
+@kb.add("ctrl-t")
+def _(event):
+    "When ControlT has been pressed, toggle visibility."
+    hidden[0] = not hidden[0]
+    print("\nType Control-T to toggle password visible.")
+
+Toggle = Condition(lambda: hidden[0]) #,bind=kb